refactor: Split models and handlers into domain-specific files
- Split models.go into:
  - common.go: Common response and error models
  - project.go: Project-related models
  - folder.go: Folder-related models
  - bucket.go: Bucket-related models with advanced options
  - auth.go: Authentication models

- Split handlers.go into:
  - handler.go: Handler struct and NewHandler function
  - project_handler.go: Project CRUD handlers
  - folder_handler.go: Folder CRUD handlers
  - bucket_handler.go: Bucket CRUD handlers

- Improved code organization following Go best practices
- Maintained all Swagger annotations and documentation
- Updated Swagger docs with new file structure
- Verified compilation and linting passes
