+++ docs/swagger.json
@@ -29,7 +29,7 @@
                         "BearerAuth": []
                     }
                 ],
-                "description": "Create a new Google Cloud Storage bucket with optional advanced security and compliance features",
+                "description": "Create a new Google Cloud Storage bucket with the specified parameters including advanced options for KMS encryption, retention policies, and access controls",
                 "consumes": [
                     "application/json"
                 ],
@@ -39,7 +39,7 @@
                 "tags": [
                     "Buckets"
                 ],
-                "summary": "Create a new GCS bucket",
+                "summary": "Create a new Cloud Storage bucket",
                 "parameters": [
                     {
                         "description": "Bucket creation request",
@@ -139,6 +139,12 @@
                     },
                     "404": {
                         "description": "Not Found",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse"
+                        }
+                    },
+                    "500": {
+                        "description": "Internal Server Error",
                         "schema": {
                             "$ref": "#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse"
                         }
@@ -313,6 +319,12 @@
                         "schema": {
                             "$ref": "#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse"
                         }
+                    },
+                    "500": {
+                        "description": "Internal Server Error",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse"
+                        }
                     }
                 }
             },
@@ -481,6 +493,12 @@
                     },
                     "404": {
                         "description": "Not Found",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse"
+                        }
+                    },
+                    "500": {
+                        "description": "Internal Server Error",
                         "schema": {
                             "$ref": "#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse"
                         }
