+++ docs/swagger.yaml
@@ -208,8 +208,9 @@ paths:
     post:
       consumes:
       - application/json
-      description: Create a new Google Cloud Storage bucket with optional advanced
-        security and compliance features
+      description: Create a new Google Cloud Storage bucket with the specified parameters
+        including advanced options for KMS encryption, retention policies, and access
+        controls
       parameters:
       - description: Bucket creation request
         in: body
@@ -239,7 +240,7 @@ paths:
             $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
       security:
       - BearerAuth: []
-      summary: Create a new GCS bucket
+      summary: Create a new Cloud Storage bucket
       tags:
       - Buckets
   /buckets/{name}:
@@ -302,6 +303,10 @@ paths:
             $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
         "404":
           description: Not Found
+          schema:
+            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
+        "500":
+          description: Internal Server Error
           schema:
             $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
       security:
@@ -408,6 +413,10 @@ paths:
           description: Not Found
           schema:
             $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
+        "500":
+          description: Internal Server Error
+          schema:
+            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
       security:
       - BearerAuth: []
       summary: Get a GCP folder by ID
@@ -510,6 +519,10 @@ paths:
             $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
         "404":
           description: Not Found
+          schema:
+            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
+        "500":
+          description: Internal Server Error
           schema:
             $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
       security:
