+++ internal/models/folder.go
@@ -1,0 +1,21 @@
+package models
+
+import "time"
+
+// FolderRequest represents a request to create a GCP folder
+type FolderRequest struct {
+	DisplayName string `json:"display_name" validate:"required,min=1,max=100" binding:"required" example:"Development Environment"`
+	ParentID    string `json:"parent_id" validate:"required,numeric" binding:"required" example:"123456789012"`
+	ParentType  string `json:"parent_type" validate:"required,oneof=organization folder" binding:"required" example:"organization"` // "organization" or "folder"
+}
+
+// FolderResponse represents a GCP folder response
+type FolderResponse struct {
+	Name        string    `json:"name"`
+	DisplayName string    `json:"display_name"`
+	ParentID    string    `json:"parent_id"`
+	ParentType  string    `json:"parent_type"`
+	State       string    `json:"state"`
+	CreateTime  time.Time `json:"create_time"`
+	UpdateTime  time.Time `json:"update_time"`
+}
